requirements.txt
absl-py==2.0.0
accelerate==0.24.1
addict==2.4.0
aliyun-python-sdk-core==2.14.0
aliyun-python-sdk-kms==2.16.2
antlr4-python3-runtime==4.9.3
black==23.11.0
boto3==1.29.6
botocore==1.32.6
cachetools==5.3.2
chumpy==0.70
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.0
contourpy==1.2.0
crcmod==1.7
cryptography==41.0.5
cycler==0.12.1
Cython==3.0.5
debugpy==1.8.0
decorator==4.4.2
diffusers==0.23.1
easydict==1.11
editdistance==0.6.2
einops==0.7.0
fairscale==0.4.4
filelock==3.13.1
filterpy==1.4.5
flatbuffers==23.5.26
fonttools==4.45.1
freetype-py==2.4.0
fsspec==2023.10.0
fvcore==0.1.5.post20221221
google-auth==2.23.4
google-auth-oauthlib==1.1.0
grpcio==1.59.3
huggingface-hub==0.19.4
humanfriendly==10.0
hydra-core==1.3.2
imageio==2.33.0
imageio-ffmpeg==0.4.9
iopath==0.1.9
ipython-genutils==0.2.0
jmespath==0.10.0
joblib==1.3.2
json-tricks==3.17.3
kiwisolver==1.4.5
lazy_loader==0.3
llvmlite==0.41.1
Markdown==3.5.1
markdown-it-py==3.0.0
matplotlib==3.8.2
mdurl==0.1.2
mmpose==0.28.0
model-index==0.1.11
moviepy==1.0.3
mpmath==1.3.0
munkres==1.1.4
mutagen==1.47.0
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.2.1
numba==0.58.1
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.16.3
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.28
ordered-set==4.1.0
oss2==2.17.0
pandas==2.1.3
pathspec==0.11.2
Pillow==10.1.0
plyfile==1.0.2
Polygon3==3.0.9.1
portalocker==2.8.2
proglog==0.1.10
protobuf==4.23.4
psutil==5.9.6
pyaes==1.6.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocoevalcap==1.2
pycocotools==2.0.7
pycryptodome==3.19.0
pycryptodomex==3.19.0
pyglet==1.5.27
PyOpenGL==3.1.0
pyparsing==3.1.1
pyrender==0.1.45
Pyrogram==2.0.106
pytesseract==0.3.10
PyYAML==6.0.1
rapidfuzz==3.5.2
regex==2023.10.3
requests-oauthlib==1.3.1
rich==13.4.2
rsa==4.9
s3transfer==0.7.0
sacremoses==0.1.1
safetensors==0.4.0
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.4
shapely==2.0.2
smplx==0.1.28
supervision==0.16.0
sympy==1.12
tabulate==0.9.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.3.0
TgCrypto==1.2.5
threadpoolctl==3.2.0
tifffile==2023.9.26
timm==0.4.12
tokenizers==0.15.0
torch==2.1.1
torchgeometry==0.1.2
torchvision==0.16.1
tqdm==4.65.2
transformers==4.35.2
trimesh==4.0.5
triton==2.1.0
tzdata==2023.3
urllib3==1.26.18
websockets==12.0
Werkzeug==3.0.1
xtcocotools==1.14.3
yacs==0.1.8
yapf==0.40.2
youtube-dl==2021.12.17
yt-dlp==2023.11.16
